pyproject.toml
[project]
name = "huntrix"
version = "0.1.0"
description = "Practice project for OpenTelemetry, Prometheus, and Grafana"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "opentelemetry-api>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
    "opentelemetry-instrumentation-fastapi>=0.49b2",
    "opentelemetry-exporter-prometheus>=0.49b2",
    "opentelemetry-exporter-otlp>=1.28.2",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = []
